requirement.txt
google-genai
streamlit
Pillow
gTTS